Wish list; what we want to work on in details?
I made small changes to the wish list I made last week. It is a detailed description of what Colette asked for and it would fit somewhere in the beginning of README. We should be able to achieve most of the goals, so it does not seem to be unrealistic. I expect more editing of the list will take place as the project will develop.